feat: 完善公招、寻访与干员资产链路

+++ app/src/main/java/com/blueskybone/arkscreen/domain/model/AppUpdateInfo.kt
@@ -9,7 +9,7 @@ package com.blueskybone.arkscreen.domain.model
  * data层直接引用这个，checkupdate的逻辑暴露在usecase里。
  * */
 data class AppUpdateInfo(
-    var versionCode: Float = 0F,
+    var versionCode: Long = 0L,
     var version: String = "",
     var date: String = "",
     var content: String = "",
