+++ app/src/main/java/com/blueskybone/arkscreen/domain/model/ResourceUpdateInfo.kt
@@ -5,7 +5,7 @@ package com.blueskybone.arkscreen.domain.model
  * Date: 2026/7/9
  */
 data class ResourceUpdateInfo(
-    var versionCode: Float = 0F,
+    var versionCode: Long = 0L,
     var version: String = "",
     var date: String = "",
     var content: String = "",
